# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "chat_backend",
    "version": "v1.0"
  },
  "paths": {
    "/chat/messages/get": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает входящие и исходящие сообщения текущего пользователя с учётом пагинации",
        "requestBody": {
          "description": "Параметры пагинации",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagingSettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PagingSettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PagingSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/messages/send": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Посылает сообщение",
        "requestBody": {
          "description": "Сообщение",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageQuery"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/messages/count": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает количество сообщений (исходящих и входящих) текущего пользователя",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message/mark-as-read": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Отмечает сообщение как прочитанное и возвращает его",
        "parameters": [
          {
            "name": "messageGuid",
            "in": "query",
            "description": "Guid сообщения",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecieverInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecieverInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecieverInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationFailure"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/online-users": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает список подключенных к чату пользователей",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message/default-templates": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает стандартные шаблоны сообщений, доступные пользователю\r\nШаблоны отсортированы по Id (по возрастанию)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultMessageTemplate"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultMessageTemplate"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DefaultMessageTemplate"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message/user-templates": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает пользовательские шаблоны сообщений\r\nШаблоны отсортированы по ModifiedDateTime (по возрастанию)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageTemplate"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageTemplate"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageTemplate"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message/placeholders": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает доступные плейсхолдеры для шаблонов сообщений",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Placeholder"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Placeholder"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Placeholder"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/chat/message/fill-placeholders": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Возвращает текст шаблона сообщения, в котором все плейсхолдеры заменены на данные исходя из\r\nконтекста посыльного",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FillPlaceholders"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FillPlaceholders"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FillPlaceholders"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMessageQuery": {
        "required": [
          "content",
          "personGuid",
          "recieversGuids"
        ],
        "type": "object",
        "properties": {
          "recieversGuids": {
            "minLength": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Список Guid получателей сообщения"
          },
          "personGuid": {
            "minLength": 1,
            "type": "string",
            "description": "Guid военнослужащего, отправляющего сообщение",
            "format": "uuid"
          },
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string",
            "description": "Текст сообщения",
            "example": "Текст сообщения."
          }
        },
        "additionalProperties": false,
        "description": "Данные для отправки сообщения"
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "guid": {
            "type": "string",
            "description": "Guid сообщения",
            "format": "uuid",
            "readOnly": true
          },
          "senderInfo": {
            "$ref": "#/components/schemas/SenderInfo"
          },
          "recieversInfo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecieverInfo"
            },
            "description": "Информация о получателях сообщения",
            "nullable": true,
            "readOnly": true
          },
          "content": {
            "type": "string",
            "description": "Текст сообщения",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Сообщение"
      },
...
      "PagingSettings": {
        "type": "object",
        "properties": {
          "pageSize": {
            "maximum": 100,
            "minimum": 5,
            "type": "integer",
            "description": "Количество сообщений, выводимых на одной странице",
            "format": "int32"
          },
          "page": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Параметры пагинации"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RecieverInfo": {
        "type": "object",
        "properties": {
          "recieverGuid": {
            "type": "string",
            "description": "Guid получателя",
            "format": "uuid",
            "readOnly": true
          },
          "recievedDateTime": {
            "type": "string",
            "description": "Дата и время получения сообщения. Default, если сообщение ещё не было получено",
            "format": "date-time",
            "readOnly": true
          },
          "readDateTime": {
            "type": "string",
            "description": "Дата и время отметки сообщения как прочтённого. Default, если сообщение ещё не было прочтено",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Информация о получателе сообщения"
      },
...
      "ValidationFailure": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "attemptedValue": {
            "nullable": true
          },
          "customState": {
            "nullable": true
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "formattedMessagePlaceholderValues": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "jwt_auth": {
        "type": "http",
        "description": "Please insert JWT with Bearer into field",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "jwt_auth": [ ]
    }
  ]
}